requirements.txt
langgraph==0.0.15
langchain-core==0.1.17
langchain-openai==0.0.2
python-dotenv==1.0.0
streamlit==1.31.1
openai==1.12.0
typing-extensions==4.9.0
pydantic==2.6.1 